helper function added in a new file
used the helper function to find the profile
