+++ soen341webapp/client/main.js
@@ -1,7 +1,6 @@
 import { Template } from 'meteor/templating';
 import { Posts } from '../lib/collections.js'; // import the "table"
 import { Likes } from '../lib/collections.js';
-import { UserData } from '../lib/collections.js'; // the collection "user info"
 import { Wants } from '../lib/collections.js';
 import { Accounts } from 'meteor/accounts-base'; // Accounts-ui takes care of password protection.
 import { Tracker } from 'meteor/tracker';
@@ -103,41 +102,6 @@ Template.addPost.events({
      }
   }
 });
-
-//Form used to add user data to the logged in user
-Template.addUserData.events({
-  'submit form': function(event, template) {
-    event.preventDefault(); // prevent page reload
-
-      // store the user input form fields into specific variables
-    var userId = Meteor.user().username;
-    var firstName = event.target.firstName.value;
-    var lastName = event.target.lastName.value;
-    var city = event.target.city.value;
-    var province= event.target.province.value;
-
-    if(firstName!="" && lastName!="" && city!="" && province!=""){
-      //added the information of the user in to the collections, userdata to later display on profile card.
-        UserData.insert({
-              userId,
-              firstName,
-              lastName,
-              city,
-              province
-          });
-          //clear form
-          event.target.reset();
-          //close modal
-          $('.modal').modal('close');
-          return false;
-
-    }
-    else {
-        alert("Please fill in all fields before you submit your want")
-    }
-
-  }
- });
 
 
 Template.posts.events({
